Create Midi Sequencer
This is a sequencer designed for a composition of Fibnonacci notes in variable timing rates (1/8, etc.). This sequencer uses a 12 note cycle to align the values with chromatic scale values. The # of elements of the scale may be adjusted, and the extent/length of the sequence can be modified by the fseries input.